Rename methods, change docstrings and README.md
I renamed the methods as follows:
- differentiate! -> jetcoeffs!
- stepsizeall -> stepsize (over an array of Taylor1 objects)
- propagate -> evaluate (extends a method existing on TaylorSeries)
- iterate! -> taylorstep!
- integrate! -> taylorinteg

Some methods include two cases, for a single variable and for a
vector of variables.

The external subroutine (user-provided) which implements the equations
of motion is assumed to be called as f(t, x), where t is the independent
variable and x corresponds to the dependent variable, or a vector whose
length is the number of dependent variables.