Use Taylor1 for the independent variable in the eqs of motion
We need to use a Taylor1 variable, of appropriate order, when the eqs of
motion are time-dependent. To do so, we pre-allocate a vector, and only
update the zeroth-order value (the actual time).

NOTE: It needs to be checked the Liapunov implementation, since `stability`
matrix was not changed. I think this is ok, since the Jacobian involves only
derivatives with respect to the dependent variables.